refactor : purchaseAmount 패키지 추가, 분리
