+++ src/test/java/lotto/model/validator/LotteryNumberValidatorTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatCode;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Stream;
-import lotto.Exceptions;
+import lotto.exception.Exceptions;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
