+++ src/test/java/lotto/model/winningNumber/BonusNumberTest.java
@@ -3,7 +3,7 @@ package lotto.model.winningNumber;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-import lotto.Exceptions;
+import lotto.exception.Exceptions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
